+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased 
+
+### Features Added
+- Added AWS CloudHSM signer
+- Added Generic PKCS11 signer
+
 ## 0.7.1
 
 ### Features Added